add all different home content
