+++ src/components/headerPages/headerOfHeaders.css
@@ -1,0 +1,1 @@
+.headerImageBackGround {}
