+++ src/index.css
@@ -21,7 +21,10 @@ html {
   box-sizing: border-box;
   list-style-type: none;
   text-decoration: none;
+  background-color: var(--primary__blue-color);
 }
+
+
 
 .primaryTextBlueColor {
   color: var(--primary__blue-color);
@@ -98,5 +101,5 @@ html {
 }
 
 .loaderSpinner {
-  background-color: #575963;
+  background-color: #171E43;
 }
