--- src/pages/AboutUs.jsx
@@ -1,111 +1,0 @@
-import React from "react";
-
-function AboutUs() {
-  return (
-    <div>
-      <h1>About Us Page</h1>
-      <p className="bg-primary">Lorem ipsum dolor sit amet consectetur adipisicing elit. Exercitationem ipsum dignissimos, omnis harum nam deleniti labore aspernatur cupiditate explicabo, beatae doloremque repudiandae consectetur hic aliquam. Debitis temporibus quae repudiandae! Temporibus?
-      Sint consequuntur nihil quis totam, facere corrupti voluptate obcaecati facilis adipisci placeat ad doloribus exercitationem cum. Deleniti neque laudantium ad, nam veniam consectetur voluptate provident deserunt vero, vel culpa reprehenderit?
-      Ex, fugiat quam. Temporibus assumenda a ex repellat facilis tempore exercitationem natus excepturi quaerat nesciunt dolore veritatis iusto, quos magnam non quasi corrupti delectus error ullam odit quod recusandae dicta.
-      Iusto magnam dolorum illo inventore delectus officiis magni neque nostrum, dolore repudiandae praesentium non nam iure! Qui minima quidem sapiente consectetur sit deserunt fugiat alias, eos saepe in debitis totam?
-      Quam, unde accusamus rerum cupiditate quasi corrupti quis nisi repudiandae cum amet hic nam nulla velit eveniet dolore sint, laborum labore optio dolorum magnam quisquam libero ad? Numquam, error nemo.
-      Placeat voluptatum iusto natus nisi fugiat deserunt, nemo ea inventore asperiores. Velit ipsum dicta quibusdam fuga dolore incidunt minima aperiam maiores perspiciatis, repudiandae ut debitis dolorum consequatur distinctio id similique.
-      Dolor mollitia eos fugiat sit incidunt, esse aspernatur qui. Iusto commodi non cupiditate quidem corporis. Praesentium, tenetur nihil unde necessitatibus repellat sequi facere exercitationem, consequuntur mollitia aliquam totam a ullam.
-      Quod reprehenderit error quidem, incidunt voluptatum nostrum praesentium maxime porro eum consequatur! Sed architecto quam iusto obcaecati natus distinctio illum repellat dolorem, perferendis aspernatur voluptatem nemo mollitia earum hic totam.
-      Maxime deserunt dolore sit omnis tempore incidunt ipsa, laudantium quos praesentium suscipit pariatur similique excepturi laborum officia eum, corrupti consequuntur illum facilis iste ullam sunt. Quia eaque consequuntur assumenda optio.
-      Id at perferendis obcaecati modi delectus, quia nesciunt ex temporibus labore. Quam eos iure autem adipisci corrupti vel sed maiores voluptates ullam accusantium officia doloremque assumenda quibusdam, dignissimos eum at?
-      Debitis at non reiciendis, reprehenderit architecto culpa magni aperiam eligendi quisquam, saepe et quam sapiente minus natus eveniet? Tempore asperiores sit repellat dicta, exercitationem itaque esse nobis minima iste beatae.
-      Modi quia nihil, non natus esse illum cumque rerum dolore officiis consectetur optio quibusdam assumenda facere eum! Earum tempore voluptatum vero soluta optio libero accusamus est veritatis, ab reiciendis alias?
-      Saepe veritatis accusamus vel nobis maiores! Ullam optio, accusantium quos dignissimos tenetur, vel exercitationem cumque ad doloribus ea nesciunt dolorem nihil provident minus facilis unde repellat labore iure temporibus quasi?
-      Ipsa tempora culpa accusamus odit suscipit ullam, voluptates sequi quidem doloribus ea voluptatibus asperiores blanditiis facilis in porro labore aspernatur illo. Est iusto vel exercitationem assumenda, molestiae consequuntur dolorum fugiat?
-      Aperiam voluptatum atque animi aliquam id nesciunt magni placeat quisquam, a error, possimus est? Eius similique quas, pariatur magni ipsa illo. Vitae, corporis? Magni, molestias eius! Aliquam nihil labore quam?
-      Accusantium perferendis earum error atque accusamus quam soluta. Quod porro, iste consectetur recusandae quos natus, quidem obcaecati veniam voluptates reprehenderit ex alias aspernatur et ipsa sed nisi sequi. Quidem, earum!
-      Aliquam incidunt, aut optio ab natus possimus deserunt accusamus quam mollitia iure excepturi voluptatibus quasi! Tempore perspiciatis assumenda magnam, in consectetur earum porro delectus praesentium ullam sapiente possimus. Fuga, nemo.
-      Quis voluptate commodi voluptatum iste itaque deserunt molestiae modi quas nostrum, corrupti neque. Blanditiis, corrupti pariatur, ea nulla quasi quidem neque incidunt nihil doloremque dolores exercitationem? Et eius modi consequuntur?
-      Eum quidem perferendis optio obcaecati deserunt a veniam, nulla, fuga laboriosam quo facere dolor consequatur, provident mollitia ducimus quas magnam aliquid cupiditate odit ex nam. Minima ducimus consequuntur eveniet atque!
-      Ullam magni ratione, dolore natus id reiciendis iure sequi, nisi aspernatur vero optio repellat, esse quo labore. Ratione enim nihil voluptatibus, quia accusamus a ipsum dolorum, maiores consequatur quas cum.
-      Rerum, voluptas? Vel incidunt voluptas esse! Dolores adipisci fuga, excepturi optio eveniet suscipit deleniti omnis maiores error eius natus culpa, nostrum tempore possimus laborum dignissimos, itaque facere cum! Incidunt, tempora.
-      Alias rem maiores magnam! Non dolorem quam totam harum doloremque consequuntur delectus odit aspernatur voluptates atque dolores accusantium, dignissimos omnis asperiores, veniam architecto a quia sequi veritatis alias debitis enim?
-      Voluptates quis fuga vero? Ullam odio sint ducimus exercitationem ipsa beatae, nobis ab facere. Praesentium, corporis harum repellat ipsa eaque incidunt ullam expedita dolor, pariatur vitae eius reprehenderit aliquid totam?
-      Deserunt quo labore ipsum aliquid obcaecati quasi reiciendis molestiae iste dolor tempora cum alias odit sint sequi, adipisci mollitia earum, dolore culpa ex veritatis! Nostrum, odit tempore. Veritatis, dolorum eveniet.
-      Ipsam voluptatum omnis aut ea qui nam totam fuga, facilis adipisci voluptas doloribus assumenda libero! Explicabo quis repudiandae non ratione recusandae numquam est impedit expedita quae. Cumque commodi est consectetur!
-      Culpa repudiandae quae vitae numquam hic earum quod, libero fugiat eius dolorem quasi, perspiciatis asperiores, neque iure. Est corporis ipsa labore. Quas, nulla corporis. Beatae in possimus odio maxime voluptatibus.
-      Beatae laborum voluptates quibusdam quos numquam dignissimos quo quidem rerum hic nulla aperiam sunt praesentium architecto incidunt natus, nam quam. Sequi minus dolor illum omnis, rerum inventore! Fuga, officia ex.
-      Quibusdam consequuntur deserunt voluptatibus culpa impedit vero reiciendis nobis cupiditate sed perferendis illo quisquam autem libero est, sapiente fuga ipsum rerum quidem eveniet non repudiandae quas at. Qui, quibusdam iure.
-      Totam eos unde earum cumque mollitia laudantium doloremque velit fugiat repudiandae ratione neque dicta quasi commodi, explicabo quia officiis dignissimos nulla culpa ipsum aliquam. Aspernatur fuga eaque reiciendis repellat vero!
-      Unde harum voluptatum sapiente repellat dicta optio hic quam, accusantium voluptatem esse quasi minima facere magni perspiciatis nisi iste ex dolorem aspernatur delectus eveniet placeat, sint maxime possimus! Soluta, accusamus.
-      Non, mollitia deleniti optio nam nulla excepturi vel suscipit, quis odio exercitationem illum quo iste aperiam! Iure aspernatur nobis pariatur, provident nihil dignissimos libero molestias praesentium rem quos adipisci ea.
-      Asperiores tempora nostrum consectetur accusamus cupiditate aut, temporibus sit, commodi quia voluptatem ex? Numquam facere quas doloremque? Libero quisquam dolor inventore, unde itaque ut tenetur ex autem veritatis rem iure!
-      Impedit earum rem rerum perferendis voluptatem, nulla labore ipsam aliquid assumenda fugiat ipsum sequi, mollitia omnis, voluptates eos. Ducimus fugit maxime nobis eveniet fugiat ipsam voluptatem cum. Cupiditate, illo accusamus.
-      Illo, aliquam id atque ex, reiciendis praesentium consequatur repellat quaerat at maiores animi, nemo accusamus numquam pariatur necessitatibus iusto tempore fuga. Consequuntur cumque earum fugiat rerum suscipit facere, hic iure.
-      Quam quae nihil, quod aliquid odit laboriosam impedit! Dolore quas est commodi perferendis quae. Quidem illo suscipit voluptas minima omnis veniam perferendis quaerat, nobis tenetur aspernatur quo error aut assumenda?
-      Nisi omnis impedit voluptate tempora dolor magnam unde tenetur consequuntur voluptatibus. Modi veritatis, et, mollitia id a maxime voluptas blanditiis iste eum hic est perferendis esse iure illo delectus in!
-      Ad sunt molestiae minus ipsam commodi laborum cum dolores quo dolorum accusantium non atque fugit laboriosam, magnam illo porro assumenda perspiciatis sequi iste vero vitae nisi! Vel perspiciatis quasi laborum?
-      Nobis dolore earum magnam aliquam eaque at aperiam nostrum, rem perspiciatis? Minus quo similique tempore quae consequuntur est fuga, ab dolorum excepturi modi ullam rerum officia autem perspiciatis doloribus quisquam.
-      Magni eos aut corporis consectetur, velit ut repellat repellendus voluptate facere perspiciatis quaerat ea nostrum minus quo ab, delectus ullam corrupti natus? Aliquid facere aliquam quod expedita dignissimos ad et?
-      Omnis tempore explicabo reprehenderit vitae sunt inventore minus provident deserunt quae atque? Voluptatibus, blanditiis? Distinctio nobis repellendus tempora, qui corrupti vel voluptas numquam iste esse doloremque. Optio alias repellat reiciendis?
-      Eligendi iusto ratione similique expedita voluptatum ipsa veritatis, suscipit, molestias beatae, cumque excepturi nihil illo blanditiis aliquam reiciendis! Illum, laudantium quasi at modi possimus ipsam aliquid nemo culpa error saepe.
-      Distinctio eius sequi voluptas amet dignissimos? Facere veniam vero tempore iusto ad architecto velit. Eos facilis mollitia fuga odit, earum delectus, eum beatae quasi sapiente, reiciendis id nobis! Aliquam, placeat!
-      Quo, soluta labore distinctio minima dignissimos id architecto consequatur consequuntur doloribus! Officiis magnam doloribus porro dolores reprehenderit provident, cupiditate cum fuga eaque, nemo impedit perferendis consequuntur voluptatibus. Nulla, vitae enim!
-      Illo eos quae provident quia. Vero laudantium cum hic, aperiam quisquam natus est maxime doloremque dicta officia autem ullam alias repudiandae quos odio vel dolores qui voluptatem, quaerat, quidem deleniti!
-      Natus, amet. Accusantium, dicta accusamus. Ex aliquam accusantium porro blanditiis incidunt inventore nisi temporibus ipsam debitis quia sunt accusamus delectus id cum, impedit alias! Fugiat quisquam nam deserunt obcaecati ea.
-      Doloremque, natus? Odit facilis labore nesciunt laudantium eum ut cum, voluptatibus fugiat illum hic vitae sapiente tempore perspiciatis cumque? Deleniti laborum voluptate rem. Quis, exercitationem impedit. Illum optio facere dolorem?
-      Quis, aspernatur tempora. Dicta quasi molestiae tempore excepturi quidem quia eum corrupti ipsam ea eveniet pariatur architecto est soluta ut harum doloribus dolores tempora iusto dolor provident quisquam, rerum nostrum.
-      Dolore quasi fugiat unde. Vitae doloribus placeat illo perspiciatis eligendi voluptas deserunt eius officia tempora autem numquam aliquid corporis, eum id accusamus sint iste necessitatibus ad ab! Deleniti, a dolor.
-      Deleniti accusantium commodi voluptatibus, possimus quae perferendis accusamus sunt voluptate odit aut molestiae rem dicta dolor saepe voluptatum tenetur eaque incidunt doloremque? Cum temporibus accusantium perspiciatis quaerat ipsam amet necessitatibus.
-      Nam, cupiditate tempora libero cumque facilis quaerat, dignissimos veritatis magnam officiis doloribus culpa fugit pariatur maiores, dicta reprehenderit laboriosam nemo consequatur. Odit, inventore? Odit officiis debitis voluptas perspiciatis iusto veniam!
-      Recusandae adipisci voluptatibus vitae! Beatae aliquid debitis accusantium nesciunt, corrupti voluptatum dignissimos fugiat? Aliquid consequatur architecto pariatur quo dolorem. Cum nihil molestiae sit possimus magnam sunt nesciunt incidunt minima ratione?
-      Ea sint et sed quod assumenda provident quidem iste exercitationem, labore, tenetur obcaecati suscipit magni illum voluptate alias ducimus dicta iusto quia quos aperiam magnam porro inventore minima? Modi, distinctio.
-      Cumque quae reprehenderit similique expedita. Voluptas ipsa maxime enim quos fugit iste, ullam harum autem, praesentium voluptate, nihil alias assumenda? Fugiat a tenetur dolores? Necessitatibus aspernatur nesciunt impedit deleniti expedita.
-      Saepe numquam blanditiis debitis voluptate. Architecto, quam corporis facilis dolores vitae numquam ab dolorem exercitationem eveniet eos dolorum porro repellendus, eius dolore officiis, velit placeat. Suscipit sunt voluptatibus architecto eaque!
-      Quasi accusantium saepe numquam! Nemo eius sint laborum quaerat error quidem totam, veniam distinctio, praesentium modi sed vero autem consequuntur in unde sapiente voluptate soluta provident eaque aspernatur doloremque ad.
-      Beatae quas eius fugiat distinctio hic laudantium laborum placeat praesentium incidunt sit corporis corrupti molestias magnam natus et qui sequi facilis deleniti, modi fugit dignissimos dolorum vitae. Soluta, vel earum.
-      Velit amet non enim rerum officiis quisquam facilis molestias, consequatur similique minus deleniti laboriosam, harum expedita architecto nemo excepturi totam illo nisi laborum fuga provident. Commodi quos dolorum architecto veniam.
-      Voluptatem, voluptatum. Quae, asperiores corporis perferendis nemo neque ipsam mollitia repellat porro eum vitae cum? Non doloremque dolorem est dolorum, numquam soluta ipsa, delectus asperiores voluptates dicta quidem dolor quae.
-      Veritatis incidunt debitis voluptas necessitatibus. Perferendis, ducimus accusantium ratione vitae voluptates totam quasi, ipsum aliquid molestias et labore harum assumenda sapiente asperiores quo molestiae quam officiis animi explicabo at illum.
-      Nemo, sequi illum delectus eligendi pariatur optio deleniti sed facilis quas animi iure alias quo accusamus excepturi. Accusamus nam blanditiis sint alias eum quidem ratione iusto nisi molestias! Nihil, nisi?
-      Facere rerum autem enim! Laudantium culpa itaque incidunt eius inventore sequi natus aliquam perspiciatis dolorem minima, possimus, repudiandae fuga, doloribus molestiae amet dolores iste quis rerum! Explicabo amet beatae optio?
-      Eaque voluptatibus eligendi nulla illo culpa minima quam quisquam. Quaerat accusamus porro iste. Quam cumque qui accusamus id ipsum aspernatur delectus adipisci, minima officiis, quae sed nesciunt veniam accusantium. Nemo?
-      Dignissimos quam velit, porro architecto error odio labore dolor nihil et harum delectus corrupti veniam cupiditate obcaecati nemo esse! Debitis amet officia dignissimos impedit eius ea accusantium voluptas! Et, qui.
-      Quis dicta recusandae soluta dolore voluptatibus, perferendis illo ullam explicabo tempora, nobis odio voluptates a? Voluptatibus iure, quis expedita dicta voluptate saepe vel explicabo, earum repellendus numquam suscipit voluptatem esse!
-      Saepe, mollitia? Quas minima maxime nisi animi itaque vero numquam id optio, fugiat ab voluptate totam alias soluta obcaecati aspernatur nulla dicta dolores officiis aliquid quibusdam architecto quam eligendi voluptas!
-      Voluptatem tenetur laboriosam adipisci est architecto incidunt voluptatibus amet labore a eos explicabo expedita ipsa maiores, quo veniam atque fugiat commodi, dolorum quibusdam minus rem. Incidunt ratione itaque necessitatibus suscipit.
-      Eligendi dolorem, porro debitis nobis eveniet magni magnam ipsam minima, incidunt quidem blanditiis ratione impedit placeat suscipit expedita sequi! Impedit cupiditate sequi id perferendis maxime provident. Debitis magni dicta officia.
-      Repudiandae, recusandae repellendus. Similique aspernatur quasi labore quos magnam voluptas natus ad dolores earum, fugit incidunt accusamus sint sunt alias magni? Modi dolor, laborum beatae quasi aliquid deleniti quisquam cum.
-      Provident beatae aperiam veniam incidunt sunt tenetur, placeat atque, minima alias, animi consequatur deserunt repellat facilis velit aliquid odit! Hic velit beatae exercitationem sequi quod, accusamus placeat quae aut odio.
-      Adipisci excepturi eius quaerat reprehenderit odio tenetur ad distinctio asperiores, voluptas quasi velit rerum. Natus dolor pariatur voluptates recusandae labore minus quisquam hic, et, placeat nostrum eaque, repudiandae facilis fugit.
-      Eius, animi! Saepe iure dolorem omnis asperiores in excepturi nam eum libero, sit eveniet optio alias ipsa cum illum suscipit amet quod corporis id! Atque numquam quis in quae tempore.
-      Laboriosam dicta incidunt at dolores impedit ea, error modi non sit! Expedita animi illum nesciunt inventore omnis, architecto voluptatum cupiditate tenetur ea aperiam laudantium numquam nam consectetur velit? Amet, quia!
-      Harum dolorem quae nam commodi eum ab repudiandae cumque ipsum minima, temporibus, accusamus facilis optio fugit minus perspiciatis magnam, et aperiam illo eius. Nisi accusamus excepturi repellat harum tempora dignissimos.
-      In ipsam exercitationem vitae, assumenda tenetur quo iste optio. Dolore reprehenderit architecto commodi sapiente nemo soluta eum atque quaerat molestias vero nesciunt consequatur sint ut ex, illo libero a necessitatibus?
-      Aperiam ipsam a quibusdam harum, quidem dicta molestias deserunt, unde velit perferendis eum minima quasi facilis beatae aut officia placeat! Asperiores consequatur sunt ipsam? Culpa consequuntur incidunt nobis itaque exercitationem?
-      Eveniet, illo sunt dolore non praesentium consequatur at totam. Deleniti quaerat voluptatum consequatur animi dolores! Hic voluptates eum illum. Qui quo unde harum magni recusandae esse veniam libero, neque iusto.
-      Repellendus, dignissimos? At, odit nostrum! Eos autem voluptates suscipit dolores magnam doloribus illo distinctio explicabo ipsa optio id voluptate sapiente, odit quisquam, accusantium praesentium debitis! Harum voluptatibus deserunt in perspiciatis.
-      Dicta ipsum et quidem eos debitis sint! Laborum, necessitatibus. Ab quia rem voluptates quod, rerum maxime inventore nemo hic enim eaque nostrum doloribus cupiditate quam illum facere modi, blanditiis quaerat.
-      Quam amet aperiam cupiditate dicta, ad laborum quas atque et molestias, consequatur assumenda expedita tenetur. Dolores ab nostrum minus ut ratione enim eum pariatur quis fugiat tempore. Sequi, dolores saepe.
-      Magni illo, aut, suscipit excepturi eum ratione officia exercitationem, blanditiis quos repellendus iste quaerat corporis debitis qui non sit aspernatur libero repellat dolorem veritatis dignissimos porro sint officiis delectus. Commodi!
-      Officiis, explicabo quasi, similique deserunt dignissimos nihil distinctio magni in iste quas, cupiditate vel expedita. Obcaecati eos aspernatur magni tenetur nemo non voluptatum aliquid cupiditate laudantium, vero totam voluptas minus?
-      Veritatis suscipit a laboriosam repellendus pariatur nam ducimus officia reiciendis eligendi quis, illo nemo voluptates odio atque nostrum eos, aspernatur inventore placeat, ipsum neque perferendis libero sint. Magnam, ipsa iure.
-      Consequuntur quasi quisquam ea doloribus! Explicabo fuga hic atque. Ea provident, praesentium impedit aperiam placeat ratione minus laborum enim illo, repellat magnam animi voluptas esse adipisci facilis excepturi nesciunt iste.
-      Sint porro necessitatibus doloribus explicabo rerum voluptatum odio sit exercitationem officiis! Corrupti, quis. Dolores officia voluptates quis excepturi eaque quia sit nesciunt veniam, eveniet ab minus dolorem, commodi autem eligendi!
-      Numquam, sunt? Doloribus mollitia earum nulla. At dolores officiis sit suscipit vel adipisci incidunt deleniti, in dignissimos consequatur molestiae beatae non veritatis officia, repudiandae exercitationem eveniet obcaecati eius corporis assumenda.
-      Eveniet nisi, aliquid doloremque porro adipisci error ex nam quibusdam, corrupti iusto accusamus dolorem perspiciatis minus fugiat repudiandae autem? Dolores officia doloribus corporis adipisci cum praesentium voluptatum odit soluta fuga!
-      Alias dignissimos autem eveniet neque. Consectetur soluta tempora optio sed nihil ipsa minus et dignissimos magnam voluptate, beatae quae, impedit cum amet voluptatem aliquid in inventore, iusto tenetur animi quaerat!
-      Distinctio magni voluptas architecto, similique dolorem officia exercitationem maxime itaque, esse delectus libero. Quasi ut ea quo a beatae laudantium, blanditiis corrupti? Facilis voluptatem temporibus quia error perspiciatis dicta vero!
-      Voluptas ducimus incidunt qui illum? Sint doloribus obcaecati explicabo? Saepe odit doloremque esse quasi, eveniet sequi labore illo tempore numquam. Voluptatem accusamus sapiente explicabo nisi delectus mollitia quisquam reprehenderit exercitationem!
-      Ipsa amet doloremque ab beatae aspernatur in laudantium tempora a dolores laboriosam sit veritatis, voluptatum incidunt minima sapiente! Repellat at quas consectetur aliquam veniam quos, consequatur voluptate animi quisquam culpa!
-      At suscipit officiis corporis dolor eveniet, ipsa eius consectetur harum molestias! Quibusdam voluptas necessitatibus distinctio quae totam eos, vitae aperiam, quam vel eaque veniam dolorum. Incidunt velit assumenda quas beatae.
-      Iure atque esse tempora debitis dolor nobis quam, sequi enim incidunt reprehenderit sit quos a veniam eaque laboriosam totam et? Cumque inventore commodi quo nam incidunt! Impedit quod maxime neque!
-      Soluta voluptatem animi illo ea quae ducimus impedit est tenetur doloribus odio, quia nihil qui quas ipsam molestias expedita nisi voluptatum ad nesciunt illum quasi autem in? Dolore, aut ducimus.
-      Tempora cumque, voluptatibus magnam cum nesciunt porro corporis dolorem dolorum laudantium nam! Possimus incidunt fugit molestiae blanditiis ullam est natus. Ab aut totam neque eos itaque amet distinctio qui deleniti.
-      Maxime libero quisquam ex iusto vitae numquam repudiandae dolore sed beatae esse ipsam dolores id neque vel voluptatum quia expedita, inventore nulla eos. Velit voluptatem esse voluptate ducimus, sapiente itaque.
-      Veniam facere expedita fugiat ullam odio quasi ut laboriosam vitae consequuntur nihil magnam similique commodi, earum nam dolorum nulla quia! Facere nisi repellat hic nulla atque modi quidem minima molestias?
-      Eaque odit repudiandae doloribus id earum debitis, nemo inventore ut saepe! Voluptatum hic natus possimus nemo eligendi, consequatur molestias? Id cum distinctio esse inventore, perferendis harum sit ad deserunt corporis?
-      Blanditiis harum dolores est fuga, reiciendis ut deleniti velit quibusdam totam nesciunt id hic eligendi quis distinctio amet minus perspiciatis minima cum inventore ab vel. Quas consequuntur nemo adipisci quasi.
-      Nemo autem molestiae rem cumque voluptatum, excepturi repellat impedit sit quasi sapiente optio totam placeat tenetur minima dolor blanditiis! Harum illo ullam perspiciatis, enim at ut explicabo! Aperiam, nesciunt veritatis.
-      Iure delectus magni accusantium laboriosam ullam officia doloremque laudantium animi. Mollitia, a vel exercitationem ipsam nobis consequatur ut corrupti numquam. Placeat ipsam eaque corrupti vel mollitia voluptatem doloremque in ipsum?</p>
-    </div>
-  );
-}
-
-export default AboutUs;
