--- src/pages/HomePage.jsx
@@ -1,11 +1,0 @@
-import React from "react";
-
-function HomePage() {
-  return (
-    <div>
-      <h1>Home Page</h1>
-    </div>
-  );
-}
-
-export default HomePage;
