--- src/pages/Services.jsx
@@ -1,343 +1,0 @@
-import React from "react";
-
-function Services() {
-  return (
-    <div>
-      <h1>Services Page</h1>
-      <p className="bg-success">
-        Lorem ipsum dolor sit amet consectetur adipisicing elit. Voluptatum,
-        doloribus a. Repellat, quos inventore soluta eius praesentium
-        temporibus, cupiditate veritatis aperiam voluptates asperiores, fugiat
-        reprehenderit eum voluptatum adipisci fugit architecto! Soluta aperiam
-        quam non, voluptatibus quis veniam, corrupti ut iste deleniti quasi
-        autem fuga saepe dolorem quos, odio consequatur tempore itaque
-        consectetur cum magni porro reprehenderit praesentium aspernatur. Ab,
-        repellendus. Eveniet quos necessitatibus, pariatur commodi aliquid
-        perspiciatis ipsa dicta. Deserunt illo enim itaque beatae recusandae
-        exercitationem nihil cumque praesentium adipisci dolorem, iste
-        cupiditate repudiandae voluptatem nemo. Iure aut praesentium ab?
-        Reiciendis magnam, tempore suscipit cum quaerat laborum expedita
-        adipisci laboriosam quisquam ipsam. Ullam ab suscipit, ex itaque,
-        facilis dolor mollitia maiores fugiat provident neque sed iste et
-        blanditiis amet laudantium! Quaerat voluptate aperiam beatae quas, quasi
-        molestiae, debitis earum recusandae dicta cumque cum. Tempore, itaque
-        sint quam officia hic deleniti doloribus, numquam velit esse fuga unde
-        neque accusamus qui quos. Iusto veritatis fugiat similique, repellendus
-        eaque quis quidem itaque! Nihil totam vitae voluptate deleniti minima
-        ut, alias fuga dignissimos repudiandae, tempora sapiente culpa neque
-        quos eaque fugiat suscipit dolorum debitis? Consectetur ducimus,
-        officiis consequatur tenetur commodi voluptate culpa voluptates!
-        Voluptas tenetur fugit similique in laborum minima suscipit excepturi
-        facilis praesentium sint. Voluptates maiores tempora dolor culpa, sit
-        totam repudiandae optio. Aut quis asperiores quam veritatis perspiciatis
-        sequi atque corporis laudantium consectetur magnam et fuga minus, facere
-        mollitia ut, reprehenderit ipsam. In, distinctio corporis? Sit, labore.
-        Quidem odit hic itaque? Fuga. Iure tenetur quod et sequi nulla animi ea
-        possimus aperiam, qui consequuntur id minima? Reprehenderit molestiae,
-        est magni facere, id, adipisci sapiente aut necessitatibus nisi culpa
-        vel dolore at! Nisi? Laudantium nihil ratione consectetur omnis
-        distinctio doloremque, aut aperiam soluta dicta! Veniam quae quod eum
-        dicta suscipit labore eius minus odit nulla vel, quas, facilis dolore,
-        assumenda ut et rem. Aperiam et voluptates unde corporis in adipisci
-        sapiente ipsam eveniet suscipit laudantium natus saepe quisquam
-        consequuntur officiis quod, itaque commodi cumque, deleniti iste
-        excepturi rem. Quos impedit veniam vel nam. Ab, omnis ipsum. Saepe
-        maxime molestiae optio illo nostrum necessitatibus cupiditate dolore
-        incidunt ex, beatae, recusandae aliquam dolores doloremque. Obcaecati
-        fugiat asperiores labore enim illum, iste aliquid? Reiciendis, commodi
-        temporibus. Corporis temporibus vero voluptas expedita nam itaque quia
-        est tempore sapiente et, laudantium blanditiis veritatis voluptatibus
-        quisquam iure, consectetur omnis architecto ipsa corrupti non vel!
-        Corrupti nobis incidunt repellendus excepturi. Porro, animi. Eum,
-        nostrum ipsam similique est harum eligendi vitae debitis quasi alias,
-        voluptate vero incidunt a rerum ducimus suscipit. Quod adipisci amet
-        quae corrupti beatae, quibusdam nulla vitae saepe? Nulla autem laborum
-        quas harum sequi ex placeat repellendus delectus. Aliquid autem aliquam
-        ea, facilis impedit rem quia placeat ab, vitae quasi fugiat ullam
-        molestias, porro sint perspiciatis ad quidem! Quae, repellendus fugiat
-        ratione numquam laudantium commodi eveniet fuga necessitatibus?
-        Recusandae, officia soluta. Laboriosam tenetur esse quis eum. Labore
-        obcaecati, ullam fugiat iste itaque omnis doloribus? Iure quos nemo
-        quasi! Provident vitae animi obcaecati repellendus. Harum exercitationem
-        laborum est atque tempore ut libero totam omnis incidunt perspiciatis
-        fuga voluptas fugit, dolores voluptatum obcaecati temporibus, a possimus
-        sapiente minus sit dolorem! Tenetur assumenda possimus pariatur illum
-        voluptatem cumque perferendis sit, deleniti consequatur exercitationem
-        illo minima omnis deserunt officiis praesentium cupiditate fugit
-        incidunt nam eligendi est, dolorum aliquid numquam. Voluptate, quidem
-        dolores! Dolorem voluptatum dolor amet, enim eligendi excepturi voluptas
-        ea? Animi, repudiandae. Qui iusto repudiandae est laborum deserunt
-        soluta aperiam nihil accusantium quis explicabo pariatur enim eius,
-        labore ut placeat rerum. Modi, corporis vitae? Assumenda magni quam odit
-        non asperiores impedit corporis cumque. Facilis, asperiores! Aliquam ab
-        debitis beatae iste corporis ipsum et minima atque! Id tempore est ab
-        quae repellat. Quasi doloribus incidunt officia nemo fuga? Sapiente
-        magni ducimus itaque iure eaque hic quam cumque nihil sunt ullam non
-        autem exercitationem nobis obcaecati, alias doloribus libero nostrum
-        dicta corrupti deserunt. Id eligendi animi, dicta quaerat cumque
-        corporis dolores. Nisi error corrupti nihil, illum ratione perspiciatis,
-        et quasi quis consequatur fuga laborum quas culpa incidunt distinctio
-        deserunt nostrum suscipit quibusdam iusto! Ducimus nesciunt dolores
-        natus exercitationem inventore facere in animi unde atque quasi cum
-        sequi repudiandae quae enim modi fugiat veniam dolorem, vero dolore
-        error! Tenetur voluptas labore soluta harum quam. At nulla nesciunt
-        velit amet placeat, illo reiciendis a tenetur commodi alias consequatur
-        voluptates rerum assumenda deleniti, magnam omnis quod deserunt ducimus
-        accusamus inventore cupiditate nostrum corrupti. Eius, sint mollitia.
-        Itaque aliquid repellat placeat aliquam minus. Reiciendis, sequi
-        voluptates soluta provident totam dignissimos sunt excepturi tenetur
-        ratione! Sit quaerat, beatae rem veritatis deserunt, ducimus totam est
-        expedita accusamus fugit eaque! Dolorem blanditiis, excepturi et
-        molestias delectus assumenda? Dolorum facilis, corrupti, laborum eum
-        deleniti quidem excepturi nulla, suscipit voluptatem animi neque sed
-        voluptas iste placeat id cumque odit saepe inventore. Nostrum. Optio
-        quibusdam qui est aperiam. Nemo ipsa ea perferendis laboriosam quo
-        nostrum, autem quod tempore dolor, earum dolorem alias temporibus
-        officia soluta voluptate consectetur possimus facere aperiam commodi ad
-        reiciendis? Laudantium expedita sapiente nemo, enim adipisci officia
-        quae aspernatur quo quas saepe laboriosam quia soluta consectetur
-        molestias repellendus! Quasi obcaecati distinctio deserunt, consequatur
-        excepturi vel doloribus sit temporibus eos ipsam. Maxime eaque minima
-        inventore porro sequi enim magni in corrupti nostrum, autem natus hic,
-        laboriosam quod totam ipsum esse perferendis fugit. Suscipit delectus
-        libero officiis provident soluta! Velit, optio doloribus. Odio hic
-        magnam recusandae reiciendis nostrum maiores temporibus praesentium
-        quaerat velit, quae doloremque non quod asperiores accusantium dolor
-        iusto quas sed, deleniti, dolore est assumenda. Corporis hic ratione
-        iste deserunt. Veniam soluta quaerat fugit commodi ex repellendus velit.
-        Laborum, quasi, at reprehenderit enim, eius aspernatur sapiente numquam
-        totam sunt hic necessitatibus sint explicabo dolor facilis asperiores
-        nesciunt ullam architecto aperiam! Aliquid qui minus atque molestiae
-        perspiciatis sed, modi aspernatur distinctio mollitia iusto perferendis
-        illum ducimus voluptatibus a consectetur! Asperiores debitis distinctio
-        at aspernatur unde libero eius perspiciatis nobis assumenda minus.
-        Nostrum repellendus sunt reprehenderit vel. Sapiente, minus corporis
-        fuga rem maiores reiciendis assumenda quidem repellendus vitae illum
-        animi explicabo ab mollitia dignissimos, fugiat, soluta dolores nisi
-        unde! Quidem, neque explicabo. Deleniti vel alias est non repudiandae
-        aspernatur qui obcaecati quibusdam veniam ullam nostrum officia culpa
-        doloribus consequatur cupiditate rerum distinctio hic, facere recusandae
-        eum, modi tempora? Magni beatae nihil quibusdam! Culpa, suscipit
-        sapiente! Provident tempora distinctio quis nulla, amet excepturi minima
-        in tenetur iure odio facilis itaque modi neque, consequatur illo nisi
-        inventore. Dolores necessitatibus voluptates et laboriosam nisi aut?
-        Incidunt a maiores deleniti laudantium vitae tempora quia. Ipsum hic
-        maiores, dolor iusto alias minima saepe! Dolorum ullam, vero sequi
-        corrupti est voluptatum reprehenderit dolor amet, animi minus, eum
-        nobis. Soluta sapiente, fugiat possimus repellendus et nesciunt.
-        Reprehenderit consequatur impedit vel veniam autem repellendus est.
-        Asperiores tempore, cumque nihil, assumenda temporibus nulla mollitia
-        sequi cupiditate laborum quam odio obcaecati cum. Dolore praesentium
-        culpa at sapiente, nobis cumque facere cupiditate non nemo ex pariatur
-        nostrum sed voluptate ut dolorem recusandae doloribus consequatur magnam
-        dicta id ab nulla. Explicabo veniam excepturi inventore! Harum explicabo
-        provident eum voluptate temporibus officiis? Laborum non tempore
-        veritatis earum tenetur aperiam dolor harum praesentium impedit
-        voluptates officiis incidunt atque doloremque distinctio quis iste,
-        numquam quos. Dicta, illum. Quam perferendis recusandae modi dolorum,
-        eum cupiditate incidunt blanditiis reprehenderit dolor laborum placeat
-        unde dolores vero excepturi enim quibusdam consectetur a. Iure
-        aspernatur debitis qui necessitatibus ad ipsa unde inventore. Neque odio
-        magnam, eveniet qui quis sit enim maxime ad quibusdam accusantium
-        recusandae consectetur id laborum blanditiis illo ipsam, quisquam
-        deleniti, impedit ut voluptates iste asperiores obcaecati. Vel, quo
-        voluptates. Consequuntur, reprehenderit dolore adipisci odio non
-        perspiciatis ipsum ad natus maxime! Inventore optio aut aliquid
-        consectetur quis rem molestiae expedita soluta, sunt nulla asperiores
-        neque similique impedit labore voluptatibus hic. Sint distinctio nisi
-        consectetur nulla tenetur, consequuntur optio, harum omnis numquam, in
-        dignissimos reprehenderit eos molestias nihil. Maiores itaque esse velit
-        eveniet quis illo aperiam porro nulla, adipisci, laudantium eius! Iste
-        molestiae ea eius voluptatibus nobis deleniti minus, vitae nam fugit.
-        Quam dicta recusandae excepturi fugit, placeat maxime veritatis atque
-        aliquam et, illo aperiam ea unde. Eveniet beatae placeat magni. Ex
-        consequuntur consectetur non aliquam vel, deleniti incidunt dolore
-        debitis maiores odio sapiente corrupti, in accusantium quidem magni
-        consequatur perspiciatis asperiores perferendis placeat eius doloremque
-        ipsam illo. Laudantium, omnis velit? A reiciendis earum quae aliquam
-        temporibus velit dicta explicabo culpa tempore dolores odit quidem magni
-        ducimus maxime veritatis suscipit, natus ad minima quos ut, pariatur
-        minus, tempora architecto iure. Error! Suscipit incidunt pariatur,
-        consequatur laboriosam natus dolor tempora ea laudantium veniam
-        recusandae possimus rem tempore delectus doloribus maxime fugit
-        temporibus. Quod fuga voluptate temporibus corrupti optio voluptates,
-        ducimus qui deserunt? Doloribus quasi eos sapiente quia voluptatem,
-        consequuntur adipisci quae incidunt quos consequatur iure aliquam eum
-        alias repellendus non, reiciendis eaque, placeat maxime? At earum vitae
-        possimus? Magnam harum officiis aliquid. Quasi quo dignissimos unde ipsa
-        facere quis sapiente officia suscipit repudiandae molestias beatae
-        aliquid odio iusto doloremque fugiat, corporis consectetur doloribus
-        itaque quia sint deleniti iure delectus. Officia, accusantium maiores.
-        Praesentium voluptatum pariatur asperiores perferendis corrupti
-        dignissimos cupiditate quaerat fugit unde dicta animi nisi voluptate at
-        nostrum distinctio, consequuntur numquam maxime veniam quasi recusandae
-        non blanditiis nihil. Sed, dolores provident. Eius repellat facere
-        expedita recusandae porro repudiandae exercitationem culpa quaerat nulla
-        voluptatem fugit, eligendi laudantium ipsam, distinctio vero, pariatur
-        explicabo quidem modi nostrum soluta. Optio maxime sequi ratione impedit
-        totam. Amet ea cum dolorum, incidunt consectetur corrupti delectus a
-        maxime autem ab beatae, impedit nesciunt expedita consequatur similique
-        molestias magni doloribus ipsa. Quos error ex labore recusandae cumque
-        molestias quia? Odio officiis illum mollitia enim hic modi quos omnis
-        odit vel dolor accusantium, reiciendis doloremque maxime asperiores est
-        eveniet, dicta distinctio vero. Optio repellat, ut necessitatibus fugiat
-        odit cupiditate fugit. Eos eius magnam rem, pariatur nulla eveniet saepe
-        qui laboriosam at distinctio. Mollitia unde ullam, repellat doloremque
-        neque aut nam quod voluptatem quidem aperiam distinctio, fuga dolore in
-        ipsa vero. Voluptatem placeat quibusdam mollitia labore impedit sint
-        perferendis deserunt harum fugiat, facilis debitis quo distinctio
-        corporis, corrupti aspernatur illum at laborum quae magni aliquam error
-        voluptate et? Esse, incidunt animi? Similique asperiores natus vero
-        praesentium quaerat, saepe officiis reiciendis at libero tenetur
-        eligendi ullam ipsam consequuntur, aut officia laudantium fugit commodi
-        possimus hic aspernatur eveniet dicta. Recusandae odio obcaecati dolor.
-        Quaerat, nesciunt sequi accusantium impedit beatae quam tempora qui
-        suscipit ipsa autem sint non magni doloribus. In asperiores modi,
-        sapiente iste excepturi neque quis laborum repellendus ipsam dicta
-        deleniti eum. Debitis veritatis rem animi iste nesciunt ratione eum sed,
-        voluptatem, laboriosam ad eveniet tenetur, magnam commodi excepturi
-        ipsam molestias aut. Similique totam provident quam beatae adipisci
-        aliquam ab aspernatur quod. Repellat, natus labore? In, quae quasi
-        voluptatibus amet, non perspiciatis, deserunt facere ea neque debitis
-        asperiores illum eos labore suscipit id eaque? Voluptatum impedit
-        aliquid distinctio earum, obcaecati ullam porro. Ex, odio, autem tempora
-        facere ducimus soluta magni officia fuga aliquam minima blanditiis,
-        maiores atque dolores recusandae pariatur alias. Doloribus, earum. Fuga
-        aperiam eos fugit. Corrupti illo maxime libero consequatur? Tenetur,
-        ipsum odit et excepturi nam ut nostrum cupiditate autem labore vero modi
-        numquam tempore, eveniet ea aliquam quisquam eos alias nulla sapiente
-        consequuntur ipsam architecto? Dolores officia possimus inventore?
-        Exercitationem repellendus dolore enim autem, nemo a id quasi culpa
-        beatae omnis dolor, minima similique at maiores consectetur rem ex
-        cupiditate velit laboriosam officia repellat vero deleniti nulla
-        consequuntur! Placeat! Quaerat, labore. Veritatis id dolores impedit
-        numquam? Dolores ex eligendi id expedita, aspernatur voluptas optio
-        fugit totam vel suscipit voluptates culpa. Facere, eligendi. Minus
-        pariatur placeat iusto, ea eius numquam! Modi reprehenderit ipsam atque
-        amet doloremque qui dolore facere velit molestiae neque dolorem
-        voluptatum nulla, nesciunt, corporis quis iure laudantium consectetur
-        error illo fugit? Nihil neque suscipit ullam dolor laborum! Dignissimos
-        rerum non ex doloribus velit vel minima libero commodi iusto laborum
-        debitis similique cumque veritatis repellat nesciunt, sapiente accusamus
-        repellendus nisi expedita temporibus consectetur deleniti quam fugit.
-        Eos, quisquam. Beatae laboriosam eligendi exercitationem debitis eaque
-        repellendus, ea alias veniam commodi aut reiciendis, nam ducimus
-        blanditiis dolorem non? Iure placeat sunt dolores laboriosam unde
-        expedita magni nisi earum velit officia? Molestias porro, sunt velit
-        iure voluptates officiis necessitatibus cupiditate deleniti? Animi
-        cumque sit nesciunt aut magni, laudantium nostrum assumenda vitae
-        reprehenderit hic asperiores blanditiis fuga explicabo odit unde quidem
-        libero. Alias, soluta doloribus. Molestias ullam ducimus tenetur
-        inventore at porro ipsam alias nulla! Soluta dolore voluptatum, ducimus
-        consectetur quos voluptatibus quod saepe beatae, quisquam doloremque
-        recusandae. Neque vero illum nobis! Accusantium et praesentium error
-        quasi consequatur quidem, laborum, debitis vero, dicta placeat
-        dignissimos molestias. Quaerat unde esse iste, fugiat numquam dolor
-        enim, incidunt natus quos optio voluptatibus necessitatibus asperiores
-        debitis. Ea nisi labore dicta ratione numquam rem aut quae tenetur alias
-        nobis suscipit magnam consequuntur, earum ab, ut odio culpa qui hic,
-        quas veniam beatae temporibus facere illo! Dolorem, id? Deserunt velit
-        excepturi perferendis sint, consectetur commodi natus ipsum in odit
-        molestiae mollitia, culpa est ipsa alias quae? Possimus reiciendis
-        eveniet deleniti eum error necessitatibus eaque et nesciunt mollitia
-        obcaecati? A quod iste sed ad voluptatum atque dolores, corporis beatae
-        saepe quidem natus, sapiente esse! Numquam vero dolores, mollitia
-        tenetur explicabo enim culpa. Impedit nihil ad quaerat, velit quo
-        labore? Voluptate numquam odio pariatur ducimus, beatae dolores autem
-        suscipit corrupti neque dicta deserunt voluptatum molestias cum maiores
-        cupiditate corporis alias distinctio omnis porro quas. Obcaecati
-        praesentium impedit id nobis delectus. Quasi ea ducimus quam molestias
-        dignissimos numquam, maxime cum accusamus in consequatur exercitationem
-        blanditiis sed obcaecati. Et repellendus, officia, fugit nemo ut ducimus
-        illum veritatis maiores, error quidem repellat exercitationem. Minus in
-        earum corrupti blanditiis maiores autem nostrum repellat ullam ex fugit
-        sit saepe, non magni quo ratione sed temporibus id praesentium, quos
-        tenetur rerum numquam ipsum soluta? Tenetur, cumque. Ratione pariatur
-        aperiam quo mollitia, illum obcaecati earum suscipit asperiores,
-        perferendis iusto nemo in modi odio architecto. Dolore sequi incidunt
-        voluptate porro aspernatur tenetur quibusdam accusantium soluta rerum,
-        saepe ipsum. Recusandae nobis dolorem eum! Neque nihil rem accusamus
-        quod iste repudiandae sequi laboriosam, totam ullam architecto tempora
-        et atque delectus vitae error ab deserunt. Natus sit sunt eos eaque
-        odio. Ex suscipit voluptas ab autem veritatis, repellat provident
-        maiores velit dolore placeat praesentium amet, omnis saepe rem eveniet
-        ipsa at. At temporibus nemo sed molestiae? Debitis ullam dicta
-        repudiandae necessitatibus! Quisquam incidunt odit dicta? Iure unde
-        suscipit sit quis tenetur doloribus maxime, culpa qui saepe cumque
-        officiis corrupti cum error magni modi sapiente expedita, ex ut in velit
-        est nemo. Distinctio aspernatur velit quia, illo ratione asperiores at,
-        nemo aliquid obcaecati sequi commodi dolor quo in impedit, minima non
-        magnam. Magnam itaque, voluptatibus ea officia quia et tempore qui
-        dignissimos! Fugit animi laborum alias eius, mollitia molestias dolor
-        quaerat esse dolorum magni velit odio quibusdam magnam sit, rem sunt
-        distinctio reprehenderit iusto harum molestiae, voluptatibus tempora?
-        Modi voluptatibus sed doloribus! Voluptate itaque nemo aliquam
-        architecto tempore magni velit inventore quia nostrum, ad esse
-        blanditiis ducimus voluptates dignissimos eligendi eaque natus! Autem
-        expedita nostrum libero modi error, cum porro! Incidunt, necessitatibus.
-        Pariatur delectus magnam praesentium labore. Enim magnam ipsam, aut
-        rerum quibusdam officiis illum sint deserunt dignissimos voluptatem
-        maiores debitis praesentium, quasi, voluptas suscipit quas qui repellat
-        molestiae fuga ipsa quod? A dignissimos quae vero? Obcaecati voluptates,
-        error laboriosam rerum ab mollitia est, similique soluta adipisci nemo
-        at sint sit consequatur voluptatem repudiandae ducimus perspiciatis?
-        Dolorem cupiditate dignissimos consequuntur odit magnam. Id doloremque
-        cumque totam molestias deleniti, eius obcaecati quae tempora ea commodi
-        sed voluptatum repudiandae et, ipsum quos minus est velit eaque.
-        Perferendis amet nobis, error quas deserunt in optio. Qui velit
-        consequatur impedit iure similique voluptate! Alias aspernatur, id
-        doloribus, recusandae officiis aliquid accusantium esse quod deleniti
-        ducimus iste voluptatem aperiam. Dolore aperiam a eligendi? Eveniet
-        error cumque dolor? Architecto iure aliquam repudiandae a fugit unde
-        porro rerum voluptates reiciendis asperiores. Nobis cupiditate dolores,
-        dolor et asperiores quisquam! Ea rerum amet, voluptatum blanditiis earum
-        eum quod consequuntur nesciunt sit? Architecto adipisci, ex in suscipit
-        id rerum non ipsam iusto harum. Fugit, fuga. Alias placeat tenetur quae
-        reiciendis officiis molestiae sint adipisci consequatur corporis illo
-        rerum, officia in ea aliquam. Asperiores beatae magni hic eos? Sunt,
-        minima? Amet laboriosam cupiditate at autem dignissimos, excepturi quod.
-        Autem recusandae corporis quaerat voluptatem deleniti reprehenderit
-        dolores ut quasi maiores? Nisi, eligendi! Ut, cum. Incidunt optio
-        mollitia voluptates eos nihil doloremque, delectus blanditiis
-        exercitationem quibusdam quasi. Quia dolorem eum delectus sed blanditiis
-        ut repellat sit asperiores assumenda obcaecati. Fugiat, ex! Recusandae
-        rem quae voluptate? Nihil facilis error cum similique. Libero nihil
-        necessitatibus quidem architecto sed repellendus quos cumque blanditiis.
-        Saepe et blanditiis ut ipsa quae dolore animi at impedit labore
-        voluptatum! Odit, nostrum doloremque. Nobis excepturi minima, voluptates
-        facere veniam similique distinctio mollitia error possimus, vel earum
-        facilis dolorum nam blanditiis iure nemo ab quisquam provident? Dicta
-        culpa tempora quo sint minima. Sapiente, officiis. Sit omnis natus,
-        repellat, ipsum officiis ea fuga dolor molestias ipsam veritatis
-        architecto suscipit velit tempora reprehenderit possimus sequi, tempore
-        eum. Sequi minima obcaecati reprehenderit quam quasi. Recusandae, id
-        eos! Neque voluptatibus earum adipisci explicabo debitis voluptatum
-        nobis veniam reprehenderit excepturi aut sit ipsam vero nostrum,
-        molestias quisquam. Nisi ipsum voluptatibus veritatis error quae magni
-        dolor iure ex cum reiciendis! Quasi cupiditate, blanditiis mollitia
-        maiores incidunt perspiciatis fugit quod odit repudiandae architecto
-        reprehenderit est, consequuntur velit ea alias nostrum necessitatibus
-        corrupti! In dolorem dolores, maxime cumque aspernatur repudiandae
-        corrupti error. Voluptatibus cum, nobis nostrum rerum hic quidem, ab
-        placeat fuga nesciunt eaque deleniti, praesentium distinctio. Dolorum
-        fugit, aut nulla praesentium tempore necessitatibus temporibus,
-        perspiciatis ex vel est, vero doloribus quia! Cum soluta eum, in commodi
-        tempore itaque enim eaque sequi numquam aut quas incidunt voluptatum
-        perferendis laborum neque expedita autem fugiat ipsum, assumenda
-        provident! Ipsa deleniti commodi tempora nobis vel. Dolor maxime
-        dignissimos velit quibusdam voluptas et voluptatum tempora cumque odio
-        eligendi possimus illo a quis, fugit in facere ipsa soluta harum
-        suscipit necessitatibus ipsam veniam temporibus atque! Sequi, laborum.
-        Architecto neque fugiat ea consectetur voluptate quas voluptatum, fugit
-        illo maiores amet hic sint ullam deserunt. Quae nostrum quasi iure
-        aliquid aspernatur beatae, asperiores quibusdam hic. Nostrum debitis
-        provident eligendi. Repellat animi, neque itaque tenetur rem laboriosam?
-        Porro asperiores vero voluptatum saepe aliquid? Ad vitae distinctio
-        blanditiis ex corporis cumque, quas facere nihil explicabo expedita
-        consectetur, debitis deserunt nostrum obcaecati.
-      </p>
-    </div>
-  );
-}
-
-export default Services;
