+++ src/pages/aboutPages/AboutUs.jsx
@@ -1,0 +1,45 @@
+import React from "react";
+
+function AboutUs() {
+  return (
+    <div>
+      <h1>About Us Page</h1>
+      <p className="bg-primary">
+        Lorem ipsum dolor sit amet consectetur, adipisicing elit. Sunt veniam
+        amet labore earum officia incidunt consequuntur saepe, deserunt aliquam
+        soluta itaque nostrum atque voluptates voluptatum fuga commodi nesciunt
+        et optio? Minus ea ab cumque minima, sapiente dolores nesciunt hic? Vel,
+        rem deserunt? Ut deleniti optio dolores. Aperiam, cum quidem eius saepe
+        ducimus beatae veniam similique magni? Quibusdam quas vel itaque! Saepe
+        natus facere nesciunt. Possimus inventore praesentium harum minus
+        consectetur accusamus a sapiente sunt enim, quod nesciunt esse nihil
+        dignissimos earum iusto quisquam natus ipsa libero non rem corporis
+        veritatis! Repudiandae dignissimos ipsum et enim, rem, aut voluptate
+        quisquam totam unde, perferendis labore recusandae blanditiis illum ea
+        nemo soluta in itaque delectus praesentium veniam. Molestias ut
+        assumenda dolore quas reiciendis. At aspernatur amet cumque rem libero
+        iste consectetur, unde excepturi iure eveniet magnam cupiditate possimus
+        quos voluptatum. Accusantium repellendus dolor ex eos? Unde adipisci
+        exercitationem atque cumque quos neque impedit! Atque dicta suscipit
+        vero ad recusandae animi asperiores voluptatum pariatur, illum
+        laudantium minima quos tempora vitae esse, accusantium perspiciatis eius
+        officia, sed ut. Voluptate eum odit quis eius ullam architecto. Totam
+        dicta, laborum adipisci odio nam modi omnis eligendi molestias autem,
+        explicabo quia sunt optio suscipit necessitatibus eius, ipsa illum
+        porro! Eveniet totam facere nisi quia libero magnam reprehenderit minus.
+        Odit enim dignissimos eveniet, cumque dolor corrupti velit,
+        exercitationem illo dolorem nesciunt earum quam ex in voluptatum eaque?
+        Quis, atque! Sapiente aliquam totam quos illum, nemo cum provident minus
+        laboriosam. Rerum illum, eum quaerat mollitia repellendus voluptates
+        perspiciatis fugit velit! Iure commodi voluptatum temporibus quisquam
+        animi quos quasi debitis illum ratione alias harum distinctio ipsa
+        adipisci neque, sit expedita autem. Provident, in? Ratione modi earum
+        dolor nemo ad blanditiis porro quos sit quisquam quibusdam, adipisci
+        libero, odio soluta quia unde. Eveniet officia earum odio et quos
+        debitis sed harum neque.
+      </p>
+    </div>
+  );
+}
+
+export default AboutUs;
