+++ src/pages/homePages/educationBackBone/education.css
@@ -1,0 +1,3 @@
+/* .ducation___container {
+    background-image: url("../../../assets/images/whiteboard-training.jpg");
+} */
