+++ src/pages/homePages/home__page.css
@@ -1,0 +1,3 @@
+.findCourses {
+    margin-top: 300rem;
+}
