+++ src/pages/homePages/whatWeOffer/whatWeOffer.css
@@ -1,0 +1,39 @@
+.iconContainer {
+    width: 50px;
+    height: 50px;
+    border-radius: 50%;
+    border: 1px solid var(--primary__white-color);
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    font-size: xx-large;
+}
+
+.contentContainer {
+    display: flex;
+    flex-direction: column;
+    gap: 1rem;
+    justify-content: center;
+    align-items: center;
+}
+
+/* .card__content {
+    background-color: var(--secondary__blue-color);
+} */
+.iconContainer {
+    width: 50px;
+    height: 50px;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+
+}
+
+.MainTittle {
+    font-size: 48px;
+    padding: 3rem 0rem;
+}
+
+.title {
+    font-size: 24px;
+}
