+++ src/pages/servicesPages/Services.jsx
@@ -1,0 +1,14 @@
+import React from "react";
+import HeaderPageComponent from "../../components/headerPages/HeaderPageComponent";
+
+function Services() {
+  return (
+    <div>
+      <HeaderPageComponent />
+      <h1>Services Page</h1>
+      <p className="bg-success"></p>
+    </div>
+  );
+}
+
+export default Services;
